Fixes #56 (https://github.com/Modernizr/modernizr-neue/issues/56)
We need to specify the file extension explicitly inside the HTML5
download attribute since the HREF attribute doesn't point to an actual
file, rather a JS blob, this trips up browsers when they are trying to
interpret the file type - even though this has been specified in the
setup.

Also fixed up grunt config file for consistency with the other two file
names.